// src/types/product.type.ts
export type Product = {
  id: string;
  images: string[];
  title: string;
  description: string;
  brand: string;
  category: string;
  gender: string;
  original_price: number;
  price: number;
  material: string;
  color_name: string;
  active: boolean;
  meta_title: string;
  meta_description: string;
  tags: string[];
  created_at: string;
};